requirements.txt
adlfs==2024.4.1
affine==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
altdss==0.2.4
annotated-types==0.7.0
attrs==23.2.0
azure-core==1.30.2
azure-datalake-store==0.0.53
azure-identity==1.17.1
azure-storage-blob==12.20.0
black==25.1.0
bokeh==3.6.0
Cartopy==0.23.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comtypes==1.4.7
contextily==1.6.0
contourpy==1.2.1
croniter==1.4.1
cryptography==42.0.8
cycler==0.12.1
dacite==1.9.2
ditto==1.2.4
ditto.py==0.2.4
dss-python==0.15.7
dss_python_backend==0.14.5
et-xmlfile==1.1.0
flake8==7.1.2
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==0.18.3
geographiclib==2.0
geojson==3.1.0
geopandas==1.0.1
geoplot==0.5.1
geopy==2.4.1
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==4.9.4
mapclassify==2.6.1
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mercantile==1.2.1
mplcursors==0.6
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==2.8.8
numpy==1.26.4
opencv-python==4.10.0.84
opendssdirect.py==0.9.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pandas-access==0.0.1
pathspec==0.12.1
pillow==10.4.0
planetary-computer==1.0.0
platformdirs==4.3.7
portalocker==2.10.0
py-dss-interface==2.0.4
pyarrow==16.1.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
PyHamcrest==2.1.0
PyJWT==2.8.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
pystac==1.10.1
pystac-client==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rasterio==1.3.10
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sns==0.1
snuggs==1.4.7
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.2
xlrd==2.0.1
XlsxWriter==3.2.0
xmltodict==0.13.0
xyzservices==2024.6.0
yarl==1.9.4
